Finally Got Fetch Working.
Yeah... I was having trouble with fetch. Been a while since I set up requests.
Forgot that there was a state management aspect to consider, that you couldn't
just return the value from within a promise chain, but must instead set it equal
to a different variable (often a state constant) that scopes the promise chain.

Making use of hooks, and creating and atply named useFetch component that will
suit most of my needs in the near term. Curious as to how I'll fold things into
Redux if/when I end up using it.