Figured out my TypeScript woes...
Well, now I know that javascript's 'push' method returns a number. Wish I'd
known that yesterday! I suppose I should thank TypeScript for the educational
experience.

I've also discovered that I am in desparate need of a deeper understanding of
Fetch requests. I think the main point of confusion was that I didn't know I
needed to access the returned value within the request lifecycle. Adding to the
frustration, it appears that I need to capture the request data within the state in
React. I think it's probably a scoping issue. It might be that I could capture
the request data in a variable that's declared outside of the fetch function.